Moved datamodel to own module
